TAS-34 moving and renaming
