+++ src/pages/ScenarioCreateEditPage/types/scenario.ts
